Rename apso-json to apso-circe
