+++ circe/src/main/scala/com/velocidi/apso/circe/JsonConvert.scala
@@ -1,4 +1,4 @@
-package com.velocidi.apso.json
+package com.velocidi.apso.circe
 
 import scala.collection.JavaConverters._
