+++ circe/src/main/scala/com/velocidi/apso/circe/package.scala
@@ -1,0 +1,5 @@
+package com.velocidi.circe
+
+/** Provides utility classes and methods related to circe-specific JSON handling.
+  */
+package object circe
