+++ circe/src/test/scala/com/velocidi/apso/circe/ExtraJsonProtocolSpec.scala
@@ -1,4 +1,4 @@
-package com.velocidi.apso.json
+package com.velocidi.apso.circe
 
 import java.net.URI
