--- json/src/main/scala/com/velocidi/apso/json/package.scala
@@ -1,5 +1,0 @@
-package com.velocidi.apso
-
-/** Provides utility classes and methods related to JSON handling.
-  */
-package object json
